+++ astropy/io/fits/column.py
@@ -19,7 +19,7 @@ from .util import (pairwise, _is_int, _convert_array, encode_ascii, cmp,
                    NotifierMixin)
 from .verify import VerifyError, VerifyWarning
 
-from ...extern.six import string_types, iteritems
+from ...extern.six import string_types, iteritems, PY2
 from ...utils import lazyproperty, isiterable, indent
 from ...utils.compat import ignored
 from ...utils.exceptions import AstropyDeprecationWarning
@@ -1429,6 +1429,12 @@ class ColDefs(NotifierMixin):
                     dt = np.dtype((dt.char + str(dim[-1]), dim[:-1]))
                 else:
                     dt = np.dtype((dt.base, dim))
+
+            # On Python 2, force the `name` to `str`
+            # because Numpy structured arrays can't handle unicode `name`
+            # See https://github.com/astropy/astropy/issues/5204
+            if PY2:  # pragma: py2
+                name = str(name)
 
             fields.append((name, dt))