+++ astropy/io/fits/tests/test_table.py
@@ -1505,6 +1505,15 @@ class TestTableFunctions(FitsTestCase):
             assert hdu.name == 'FOO'
             assert hdu.header['EXTNAME'] == 'FOO'
 
+    def test_unicode_colname(self):
+        """
+        Regression test for https://github.com/astropy/astropy/issues/5204
+        "Handle unicode FITS BinTable column names on Python 2"
+        """
+        col = fits.Column(name=u'spam', format='E', array=[42.])
+        # This used to raise a TypeError, now it works
+        fits.BinTableHDU.from_columns([col])
+
     def test_bin_table_with_logical_array(self):
         c1 = fits.Column(name='flag', format='2L',
                          array=[[True, False], [False, True]])